chore: extract save images to the image package
