+++ docker.go
@@ -1734,31 +1734,10 @@ func (p *DockerProvider) ListImages(ctx context.Context) ([]ImageInfo, error) {
 	return tcimage.List(ctx)
 }
 
+// Deprecated: use testcontainers-go [image.SaveToTar] instead
 // SaveImages exports a list of images as an uncompressed tar
 func (p *DockerProvider) SaveImages(ctx context.Context, output string, images ...string) error {
-	outputFile, err := os.Create(output)
-	if err != nil {
-		return fmt.Errorf("opening output file %w", err)
-	}
-	defer func() {
-		_ = outputFile.Close()
-	}()
-
-	imageReader, err := p.client.ImageSave(ctx, images)
-	if err != nil {
-		return fmt.Errorf("saving images %w", err)
-	}
-	defer func() {
-		_ = imageReader.Close()
-	}()
-
-	// Attempt optimized readFrom, implemented in linux
-	_, err = outputFile.ReadFrom(imageReader)
-	if err != nil {
-		return fmt.Errorf("writing images to output %w", err)
-	}
-
-	return nil
+	return tcimage.SaveToTar(ctx, output, images...)
 }
 
 // PullImage pulls image from registry
