+++ modules/k3s/k3s.go
@@ -13,6 +13,7 @@ import (
 	"gopkg.in/yaml.v3"
 
 	"github.com/testcontainers/testcontainers-go"
+	"github.com/testcontainers/testcontainers-go/image"
 	"github.com/testcontainers/testcontainers-go/log"
 	"github.com/testcontainers/testcontainers-go/wait"
 )
@@ -202,11 +203,6 @@ func unmarshal(bytes []byte) (*KubeConfigValue, error) {
 
 // LoadImages loads images into the k3s container.
 func (c *K3sContainer) LoadImages(ctx context.Context, images ...string) error {
-	provider, err := testcontainers.ProviderDocker.GetProvider()
-	if err != nil {
-		return fmt.Errorf("getting docker provider %w", err)
-	}
-
 	// save image
 	imagesTar, err := os.CreateTemp(os.TempDir(), "images*.tar")
 	if err != nil {
@@ -216,7 +212,7 @@ func (c *K3sContainer) LoadImages(ctx context.Context, images ...string) error {
 		_ = os.Remove(imagesTar.Name())
 	}()
 
-	err = provider.SaveImages(context.Background(), imagesTar.Name(), images...)
+	err = image.SaveToTar(context.Background(), imagesTar.Name(), images...)
 	if err != nil {
 		return fmt.Errorf("saving images %w", err)
 	}
